% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancycut.R
\name{fancycut}
\alias{fancycut}
\title{Like \code{cut}, turn a vector of numbers into a factor}
\usage{
fancycut(x, na.bucket = NA, unmatched.bucket = NA, out.as.factor = TRUE,
  ...)
}
\arguments{
\item{x}{a numeric vector}

\item{na.bucket}{what level should NA values be given?}

\item{unmatched.bucket}{what level should numbers not covered by an interval be given?}

\item{out.as.factor}{default is TRUE
Should the resulting vector be a factor?
If FALSE will return a character vector.}

\item{...}{These take the form \code{tag = value}.
Tags become the bucket names and values the interval definitions.}
}
\description{
Like \code{cut}, turn a vector of numbers into a factor
}
\examples{
fancycut(
  x = -10:10,
  Zero = 0,
  Small = '(0,2]',
  Medium = '(2,5)',
  Large = '[5,10]'
)
}
